requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via redis
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   redis-om
    #   uvicorn
cryptography==40.0.2
    # via
    #   types-pyopenssl
    #   types-redis
fastapi==0.95.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
hiredis==2.2.2
    # via redis-om
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
more-itertools==8.14.0
    # via redis-om
pptree==3.1
    # via redis-om
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via
    #   fastapi
    #   redis-om
pyjwt==2.6.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
python-ulid==1.1.0
    # via redis-om
pyyaml==6.0
    # via uvicorn
redis==4.5.4
    # via
    #   -r requirements.in
    #   redis-om
redis-om==0.1.2
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.26.1
    # via fastapi
types-pyopenssl==23.1.0.2
    # via types-redis
types-redis==4.5.4.1
    # via redis-om
typing-extensions==4.5.0
    # via
    #   pydantic
    #   redis-om
urllib3==1.26.15
    # via requests
uvicorn[standard]==0.21.1
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.2
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
